requirements.txt
Flask
Celery
redis
flask_sqlalchemy
python-dotenv
flask_marshmallow
marshmallow-sqlalchemy
matplotlib
marshmallow
psycopg2-binary
bs4
requests
setuptools>=68.0.0 
numpy>=1.2
pandas==2.0
scipy==1.10
sympy==1.11
tqdm==4.65
gdown==4.0.0
debugpy
gunicorn
loky
imageio
billiard
flask-cors